Organize todos by user

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserServerSession } from "@/auth/actions/auth-actions";
 import prisma from "@/lib/prisma";
 import { NextRequest, NextResponse } from "next/server";
 import * as yup from "yup";
@@ -33,10 +34,18 @@ const postSchema = yup.object({
 });
 
 export async function POST(request: NextRequest) {
+  const user = await getUserServerSession();
+  if (!user) {
+    return NextResponse.json("No autorizado", { status: 401 });
+  }
   try {
-    const body = await postSchema.validate(await request.json());
+    const { complete, description } = await postSchema.validate(
+      await request.json()
+    );
 
-    const todo = await prisma.todo.create({ data: body });
+    const todo = await prisma.todo.create({
+      data: { complete, description, userId: user.id },
+    });
 
     return NextResponse.json(todo);
   } catch (error) {
@@ -45,10 +54,17 @@ export async function POST(request: NextRequest) {
 }
 
 export async function DELETE(request: NextRequest) {
+  const user = await getUserServerSession();
+
+  if (!user) {
+    return NextResponse.json("No autorizado", { status: 401 });
+  }
+
   try {
     await prisma.todo.deleteMany({
       where: {
         complete: true,
+        userId: user.id,
       },
     });
